Update Components/Modal
Deleted and restructed .jsx components to start over. Starting with Modal components for now.

Got the modal frame and it's children to resize when the frames border is dragged.

In the middle of figuring out how to limit how bix a modal can be resized.

Need to finish Toolbar and display area.